Add polymorphic belongs-to support (#8)
- Always set owner on cached records
- It appears Ember Model (always the crumby citizen) does not clear its
cache between test runs. This results in lookups of cached items from
previous test runs if using the same ID. However those cached records
have an owner property that is not the current application instance!
That's very bad so in this work around we are just ensuring that the
current owner is assigned.